More refactoring: NestedRoutes and RoutingContainers
