+++ src/App.jsx
@@ -1,132 +1,46 @@
 import React, { useEffect, useState } from 'react'
-import { Link, Redirect, Route, Router, Switch, useLocation, useRouter } from 'wouter'
+import { Link, Redirect, Route, Switch, useLocation } from 'wouter'
 import AnnouncementToast from './components/AnnouncementToast'
-import Form from './components/ApplicationForm'
-import Navbar from './components/Navbar'
-import Page from './components/Page'
 import { A } from './components/Typography'
 import Landing from './containers/Landing'
 import {
-  Application,
   ApplicationConfirmation,
-  ApplicationForm,
   ApplicationReview,
   Charcuterie,
   Faq,
-  Gallery,
   Home,
   Judging,
   JudgingAdmin,
-  JudgingView,
   Livestream,
   Login,
   NotFound,
-  ProjectView,
   Schedule,
   Sponsors,
   Submission,
 } from './pages'
 import GlobalStyle from './theme/GlobalStyle'
 import ThemeProvider from './theme/ThemeProvider'
-import { AuthProvider, useAuth } from './utility/Auth'
-import {
-  APPLICATION_STATUS,
-  DB_COLLECTION,
-  DB_HACKATHON,
-  IS_DEVICE_IOS,
-  VALID_HACKATHONS,
-} from './utility/Constants'
-import HackathonProvider, { useHackathon } from './utility/HackathonProvider'
-import { HackerApplicationProvider, useHackerApplication } from './utility/HackerApplicationContext'
+import { AuthProvider } from './utility/Auth'
+import { DB_COLLECTION, DB_HACKATHON, IS_DEVICE_IOS } from './utility/Constants'
+import HackathonProvider from './utility/HackathonProvider'
+import { HackerApplicationProvider } from './utility/HackerApplicationContext'
 import {
   AdminAuthPageRoute,
   ApplicationInProgressRoute,
   AuthPageRoute,
+  NestedRoutes,
   NoAuthRoute,
   PageRoute,
 } from './utility/Routes'
+import {
+  ApplicationDashboardRoutingContainer,
+  ApplicationFormContainer,
+  GalleryContainer,
+  JudgingViewContainer,
+  ProjectViewContainer,
+} from './utility/RoutingContainers'
 import { db, getAnnouncement } from './utility/firebase'
 import notifications from './utility/notifications'
-
-const NavbarSaveOnLogout = ({ name, handleLogout }) => {
-  const { forceSave } = useHackerApplication()
-  const logout = async () => {
-    await forceSave()
-    handleLogout()
-  }
-
-  return <Navbar name={name} handleLogout={logout} />
-}
-
-const ApplicationFormContainer = ({ part }) => {
-  const { isAuthed, user, logout } = useAuth()
-  const { application } = useHackerApplication()
-
-  return isAuthed && application.status.applicationStatus === APPLICATION_STATUS.inProgress ? (
-    <>
-      <NavbarSaveOnLogout name={user.displayName} handleLogout={logout} />
-      <Form>
-        <ApplicationForm part={part} />
-      </Form>
-    </>
-  ) : (
-    <Redirect to="~/login" />
-  )
-}
-
-const JudgingViewContainer = ({ params }) => {
-  const { isAuthed } = useAuth()
-
-  return isAuthed ? (
-    <Page>
-      <JudgingView id={params.id} />
-    </Page>
-  ) : (
-    <Redirect to="~/login" />
-  )
-}
-
-const GalleryContainer = ({ params }) => (
-  <Page>
-    <Gallery />
-  </Page>
-)
-
-const ProjectViewContainer = ({ params }) => (
-  <Page>
-    <ProjectView pid={params.id} />
-  </Page>
-)
-
-const ApplicationDashboardRoutingContainer = () => {
-  const { isAuthed } = useAuth()
-  return isAuthed ? <Application /> : <Redirect to="~/login" />
-}
-
-const NestedRoutes = props => {
-  const router = useRouter()
-  const [location] = useLocation()
-  const { activeHackathon, setActiveHackathon } = useHackathon()
-  const hackathonFromURL = props.base.split('/')[2].toLowerCase()
-
-  useEffect(() => {
-    if (VALID_HACKATHONS.includes(hackathonFromURL) && hackathonFromURL !== activeHackathon) {
-      setActiveHackathon(hackathonFromURL)
-    }
-  }, [props.base, activeHackathon, setActiveHackathon])
-
-  if (!location.startsWith(props.base)) return null
-
-  if (!VALID_HACKATHONS.includes(hackathonFromURL)) {
-    return <Redirect to="~/404" />
-  }
-
-  return (
-    <Router base={props.base} key={props.base} parent={router}>
-      {props.children}
-    </Router>
-  )
-}
 
 function App() {
   const [announcementText, setAnnouncementText] = useState('')
